make vis.R work stand-alone for plotting crash-per-count-loc
vis.R previously relied on data generated by a query on a Postgres
database that has a copy of bike count and crash summary data.
that query was moved to the old/ directory in the last commit.
the R code in vis.R now does the tallying, tabulation, aggregating,
and averaging work needed to render the plot.

this also means that it reads and parses all of the bike count
data files, so now we have an example of that happening in R.